% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{material_properties}
\alias{material_properties}
\title{Standard Material Properties for Building Simulation}
\format{
A data frame with materials and columns for properties.
}
\usage{
material_properties
}
\description{
A dataset containing typical thermal properties for common building materials.
}
\keyword{datasets}
