% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_wall_temp.R
\name{simulate_wall_temp}
\alias{simulate_wall_temp}
\title{Simulate Temperature Profile Through a Multi-Layer Wall}
\usage{
simulate_wall_temp(
  materials,
  initial_temp_in,
  external_temp,
  total_time,
  dt,
  dx
)
}
\arguments{
\item{materials}{A data frame defining material properties (thickness, condauctivity, density, specific_heat).}

\item{initial_temp_in}{Initial internal temperature (Celsius).}

\item{external_temp}{External temperature (Celsius, can be a single value or a time series).}

\item{total_time}{Total simulation time (hours).}

\item{dt}{Time step (hours).}

\item{dx}{Space step (meters).}
}
\value{
A data frame with time and internal temperature values.
}
\description{
Solves the 1D heat equation using a finite difference method for a composite wall structure.
}
